Mobile pronto

+++ Back-End/sp_medical_group/sp_medical_group/Domains/TipoUsuario.cs
@@ -12,7 +12,7 @@ namespace sp_medical_group.Domains
             Usuarios = new HashSet<Usuario>();
         }
 
-        public byte IdTipoUsuario { get; set; }
+        public int IdTipoUsuario { get; set; }
         public string TituloTipoUsuario { get; set; }
 
         public virtual ICollection<Usuario> Usuarios { get; set; }
